Spell casting system, part 1
Added a lot in order to start work on spell system.  Will eventually be robust enough that spells can influence other spells cooldowns, etc.  Added enumerators to distinguish type of spell being cast and its school (fire, ice, etc.).  Added macros to the player character to save on space.  Certain things seem to be difficult to add to macros which needs to be examined further (shouldn't be the case).