Giga Refactoring (of Hell)
Remade spell system so that it works better in the long run.  Aiming needs to be readded, as well as all the rest of the spells that I deleted but this was necessary.  Now all abilities belong to a super class, Spells, which currently has one subclass, Projectiles, with one subsubclass, FireBlast, inside.